test(widget): move the widget tests into :feature:widget with their subject
All 59 unit tests for the widgets stayed in `app/src/test` when their subject
became `:feature:widget`, leaving the module's own `check` running nothing.
Two of them were wrong in ways only the move exposed.

`PrayerTrackerWidgetDataSourceTest` mocked `PrayerDao`, so it stopped compiling
the moment the data source took `PrayerRepository` — the layering fix this PR
already made. Rewritten against the repository, and two assertions got stronger
for free: `PrayerStatus.entries` now enumerates the five non-prayed statuses the
stringly-typed version listed by hand and got wrong (`LATE` and `NOT_PRAYED`
were missing), and the "unknown prayer name" case became `SUNRISE` — a real
sixth `PrayerName` that a careless `PrayerName.entries` in the data source
would let the widget count, where the old `"tahajjud"` literal can no longer
occur at all.

`WidgetGlyphGuardTest` scanned a module-relative path that ceased to exist in
`:app`. Its only floor was `dir.isDirectory`, which passes on the day the
directory is empty and fails only on the day it is gone — the same shape as the
`MaterialTextFieldGuardTest` bug earlier in this epic. It now asserts on files
actually read, verified by three negative runs: stale root, under-populated
root, and an injected glyph.

`AnalyticsReachabilityTest` scanned the old widget path too. It fails loudly
rather than narrowing silently, because the missing-root filter became an
assertion in PR 12 — repointed at the module. It reads three modules' sources
off disk, so `app/build.gradle.kts` now declares them with `inputs.dir(...)`;
without that `testDebugUnitTest` stays UP-TO-DATE when they change and the
assertion never runs, which is what hid a broken `:core:common` assertion
through two full local sweeps.

Also corrects the string-split comment in the module's build file, which was
counted from memory. Verified: the module references 33 string resources across
Kotlin, `res/xml` and its manifest, declares 17, and resolves 16 from
`:core:ui`. Five of those 16 have no consumer outside the module but stay
deliberately — `khatam_widget_description` is read by `WidgetsScreen` and
cannot move, so relocating the rest would split one set of six provider
descriptions across two modules to no end.

Test accounting: `:app` 1077 -> 1018, `:feature:widget` 59. Total unchanged at
2413 across nine modules.
